ui: Optionally display sources alongside decompiled code
This uses language-c to parse C files so that we can find the necessary
functions. It pretty prints them from the parsed AST, which unfortunately loses
comments.

Another drawback: it requires preprocessed source as inputs (so that it doesn't
have to invoke gcc/clang with mysterious cross-compilation arguments).  This is
suitable for demonstration purposes.