Fix AArch32 classification failures that broke the verifier
The changes were in macaw (but the other submodules are updated here so that we
don't drift too far).  There was a slight API change in macaw that is accounted
for in this commit.

The underlying cause (in macaw-aarch32) was that some simplification rules were
not firing aggressively enough; when multiple rules applied, only the first was
used.  The change applies the simplification rules until saturation.