Add timeouts to SAT checks
Some are allowed to fail relatively quickly, while the rest can't really be
allowed to fail without failing the whole verification process.  We use 5
seconds for the ones that are expected to fail sometimes, while we use 5 minutes
for the rest.

These will be made configurable in a future commit.