Refactor to use the offline backend (#46)
This replaces a persistent solver connection with a separate solver instance per SMT query. This will eventually allow us to verify goals in parallel. To support this, we changed how we represent symbolic state (with fresh variables instead of what4 BoundVars). This avoids a dependency on an API that only existed for online solver backends. This also enables us to have different timeouts for compulsory goals vs optional goals.

Co-authored-by: Daniel Matichuk <dmatichuk@galois.com>