Use verification hints to guide code discovery (#56)
This commit adds command line options to accept filenames for hint files in
various formats.  It parses the files (if any) to construct hints.  Hints are
threaded through to the code discovery phase, which now uses them to guide code
discovery and also to double check them (and emit warnings if an error is
detected).

Code discovery is now parallelized, since we may need to run it as many as three
times.