Make the solver and timeouts configurable from the command line
This change turns the solver and timeouts into parameters in the verifier
configuration.  As part of this change, the configuration data types have been
extracted into a separate module (Pate.Config).  Pate.Arch, Pate.Timeout, and
Pate.Solver are also new.

Most of the updates are just import/qualifier related.

With this change, Yices, CVC4, and Z3 are supported.

Fixes #48
Fixes #52