Render the proof structure in the interactive visualizer
The verifier now streams out proof results as they are computed.  The
interactive visualizer now processes those messages and builds an index over the
proof structure, rendering it on-demand.  The user has to explicitly take a
snapshot of the proof state that they want to visualize to avoid the graph from
moving while they are inspecting it.

This change adds the required javascript as submodules.

Slices of the original and patched program are now visualized as CFG slices
using a graph rendering library.  The proof structure is rendered using the same
graph rendering library.  Each node in the proof can be inspected for additional
information.  The Triple nodes (representing pre- and post-conditions) are the
most interesting for understanding the structure of the proof.  Each Triple node
is annotated with the result of attempting to verify it. Successful nodes are
green, while failures are red.  Verification failures display the counterexample
and the computed conditional equivalence predicate (if any).