Improve graph layout, remove old block rendering
Use dagre for the graph layout, as it is specialized do DAGs. It produces much
better results than the breadthfirst layout.

Also remove the old linear block rendering, as it is hard to follow and just
distracts. All of the same information (and more) is now available in the
graphical rendering.