ui: Make the block pair detail verification result optional
We want to reuse this in several contexts, some of which don't have this type of
result available. Before we were making up fake ones, but it is better to leave
it optional.

An alternative that might be even better would be to not include the result in
this rendering and render the result outside of here in the case where we have it.