Replace a call to `error`
This was triggerable when taking a snapshot of intermediate proof
states. Instead of erroring out, now we just render a "pending" node. The error
seemed to get caught in a way that prevented further progress, which was
problematic; I think it was down to some state updates being lost and
unrecoverable (since they are just streamed out - they could probably get
dropped when read but before processed if the exception triggered).