WIP
This commit adds some extra tracing and exposes some metadata about calls in the
UI for easier rendering/tracking.  The tracing should ultimately go away.

As noted in the code, it looks like mismatches in the link register (when
calling external functions that we don't analyze) cause mismatches. We could
explicitly exclude it (like the IP), but we might want to be strategic about
when we do that.