Support a parameterized method for specifying preconditions for missing functions
Previously, we assumed that the entire machine state was required to be
equivalent at the start of missing functions (both system calls and functions in
shared libraries).  That is a safely conservative approach, but it has
unfortunate consequences: strict equivalence on every program location
propagates from external function calls to all of the code before them.  This
causes spurious failures, but also interferes with computing conditional
equivalence (and additionally triggers some unfortunate non-confluent behavior
in that code).

To alleviate this issue, this commit introduces a configurable way to specify
the domains of external functions: the `ExternalDomain` type (which wraps a
function).  Right now, this is used to specify the calling conventions for
PowerPC and ARM for both external user-level calls and system calls. This is
unsound, as it does not account for accesses to the memory pointed to by
pointers passed in to functions. Moreover, the external calls assume ABI
compliance, which is true for functions compiled by compilers, but not strictly
guaranteed.

However, it achieves the immediate need and is extensible later. For example, it would
be ideal if we inspected the call target and applied appropriate precise
summaries when possible.

Note that this commit is unfortunately tangled up with some code motion that was
required to break import cycles.  Some code was factored out into new modules
and some imports were renamed, but that is the bulk of the changes (aside from
the introduction of the `ExternalDomain` and its plumbing through from the
frontend).