WIP
This change doesn't compile. This is trying to make the core of the verifier
agnostic to the binary format (currently fixed as ELF). This was almost working,
except for the complex code that handles the PowerPC TOC. That logic will really
need to be refactored first.

The explicit reliance on the TOC means that we actually have to deal with the
`binFmt` type parameter somehow, which really complicates any change like
this. It is currently handled by fixing that parameter to ELF everywhere.