Remove some uses of exceptions for control flow
The primary change here is that `checkSatisfiableWithModel` now returns an
Either value that captures any synchronous exceptions that are thrown by the
solver (due to error or timeout).  The `isPredTrue` and `isPredSat` functions
have been adapted to account for this change and choose the safe default given
their intended semantics.  They no longer bubble up exceptions (modulo async
exceptions).  These seem to be safe adaptations according to the tests and some
manual inspection of larger examples.

The other use sites of `checkSatisfiableWithModel` have been updated to
explicitly handle the solver error case in ways that are more suitable for each
context (usually returning an explicit failure node).

Fixes #103 (mostly)

Note that there is still an instance of this in `discoverPairs`, but that one
doesn't seem to be immediately problematic. This fixes the observed failures in
`checkCasesTotal`, and definitely improves results by not swallowing unproved
nodes that were getting interrupted by exceptions.