Relax the constraints we use for the AArch32 ABI
The original implementation of the AArch32 ABI (used for missing function
bodies) included r11, which is not at all accurate. However, it was necessary to
get some important examples to run.

Now that #103 is fixed by #120, this is happily no longer necessary, so our ABI
spec can now be reasonable.