[WARNING: incomplete] addresses issue #95
This is a quick-and-dirty patch that tackles issue #95. Previously, the
precise memory model would combine bytes that we read from memory via
shifting and masking. Now we select out the appropriate byte and
concatenate them together.

This patch is incomplete, because it includes an undefined in an
impossible branch. It's also aesthetically unpleasing; I'd like to
stick the selection and concatenation logic into its own function to
increase readability of readMemArr. However, because preliminary tests
suggest that this patch slows things down rather than speeding them up,
I am punting on those changes for now.