Refactor the handling of the Table of Contents on PPC64 (#126)
Instead of directly mentioning the TOC, abstract the validity checks for special
dedicated registers out into a separate abstraction.  This has the advantage of
making this code more general so that we can handle dedicated registers on other
architectures (like the Small Data Area pointer in ppc32), while also letting us
remove some hard PPC64 constraints from the core.

This includes fixes to the ppc32 loader to no longer require a TOC. With this
change, ppc32 support in pate works (but still needs some tests).

Note that there are some failing tests that are marked as expected for now until
the issue is fixed in macaw.