Fix a bug introduced by the recent TOC handling changes
The bug was not specific to the TOC handling, but that commit generalized the
`lookupBlocks` function into `lookupBlocks'`, which was usable from more
contexts. Unfortunately, that change did not work as expected in cases where an
exception was thrown. The change turned some failures into an IO exception,
whereas before they were in the ExceptT monad (and caught).

Note that the case where this occurs may benefit from more explicit handling,
but the change in this commit restores the previous behavior (and removes the IO
exception from `lookupBlocks'`.