Add support for separate hints files for original and patched binaries
Before, we only supported hints for the original binary. It is important that we
have hints for both, especially in cases, with substantial differences between
the two binaries.

This change modifies the command line arguments to support the extra specs. It
also plumbs both sets of hints through the core and uses them during code
discovery.

They are not yet used to provide names to arguments/locals, but this change
enables those names to be accurately provided to both the original and patched
binaries.

Fixes #138