Add some ad-hoc tracing for the simplifier
Print out an error to the console if the simplifier validation query proves that
the simplifier is broken.

Also print out warnings if it hits a timeout.

Previously, this code threw an exception in the case of a timeout *or* an actual
failure, making it difficult to tell the difference. Now we have more granular
information and can use that to fix failures (i.e., incorrect simplifications).

Note that this currently uses `traceM`; it might be nice to use the normal
logging facility instead.