Simplify most uses of the `ParsedFunctionMap`.
Previously the `parsedBlocksContaining` function was actually
computing all the blocks in a given function that were forward
of a given address. Here we rename this function to correctly
indicate what it computes.  Moreover, we notice that most uses
of this function should probably acually just return ALL blocks
in the body of the function.

Thus, we modify `lookupBlocks` so that it returns all blocks in
a function, except for one call site in Symbolic Execution.
Changing the behavior at this call site breaks pretty much all
of the test suite.

I conclude that the "all blocks forward of an address" lookup is
important for the loop-breaking heuristics in `simulate`, and is
otherwise probably just incorrect.