Discover code on-demand
Previously, we used macaw to eagerly run code discovery and analyze all of the
code that we could find. This was problematic for large binaries, for which we
ran out of memory. The memory expansion of macaw IR (and Crucible) compared to
machine code is enormous (probably over 1000x).

With this change, we now only run code discovery on demand (i.e., when we find a
call to a function, we analyze just that function).  We cache analyzed
functions for easy access.  Note that cache currently only grows; it could
trivially become an LRU cache if desired.

An unfortunate side effect of this change is that we can do a bit less in the
way of diagnostics. Before, we were able to figure out which hints from users
were incorrect (i.e., inconsistent with analysis results). Now that we only
lazily find code, we can't do that anymore. Some of our diagnostics and metrics
are less precise now.

Closes #155