WIP Plumb through equated function metadata
Also stub out the location to check to see if we should apply the function
inline heuristic, with some comments about what the code around that is doing.

The next step is to construct the necessary CFGs in `inlineCallee` to drive
symbolic execution and ultimately construct the `StatePred` that we want here.