Change the symbolic backend to an Online instead of Simple backend
This is a bit unfortunate, but we need the switch because we need the online
backend for the callee inlining feature, which needs path satisfiability
checking. Unfortunately, we can't just allocate a local online backend for that,
as we wouldn't be able to import or export terms between the two backends.

Now, we just allocate an online backend up-front.  Note that we still only use
the online features for the inline call feature. For the rest of the verifier,
we still issue SMT queries using the what4 `SolverAdapter` interface.