Declare appropriate problem features for the online solver
The online solver can't do any formula inspection to turn on features because it
doesn't know what formulas it will get when the connection is
opened. Pre-declare what we need, particularly symbolic arrays, which are not a
default for the yices backend.

This change also adds an option for the user to specify a solver interaction
file to record SMT solver sessions to for debugging purposes.