Populate the symbol table use for the InlineCallee feature
This includes (at least for ARM) the addresses of PLT stubs so that we can add
overrides for functions called via PLT.

This is structured a little oddly because we don't want any references to ELF
files in the core of the verifier, but parsing out PLT stub information requires
deep inspection of ELF files (and is very specific to the ELF format). Thus, we
do the parsing "outside" of the verifier when we set up all of the arch-specific
information in the SomeValidArch structure).

This change also adds a bit more structure to the `SomeValidArch` structure so
that adding fields doesn't break pattern matching everywhere anymore.