Update submodules
We needed to do this so that we can fix a bug in macaw-symbolic. Unfortunately,
bringing that up to date requires updating most of the rest of the dependencies,
too.

This doesn't yet fix the macaw-symbolic bug (memory initialization in the
presence of relocations), but this is a necessary first step.