Treat relocation values as fully symbolic during memory setup
When setting up memory for crucible, we need some initial value to use for
relocations. We don't have the values they point to (because they are in shared
libraries). We treat them as fully symbolic bytes and hope for the best (any
references will need to be overridden).

This uses a new feature in macaw-symbolic