Initialize separate memories for the two programs
Since the data and text sections do not necessarily align, we can't use the same
memory (or the patched binary will have an inconsistent memory image).  This
seems like it should be okay because we just need to prove equivalence of the
memory post-states (and only of the .data sections).  The initial values are not
very important (since we only care about the locations that have *different*
contents.

We could run into trouble if symbolic bytes are moved around that have no shared
provenance - in that case, we may need to take care to share the initial bytes
of the .data sections (but note that we still have to be careful that they *may*
not align entirely.