Update macaw to preserve block ids on some pointer operations
One major issue was that we were generating invalid memory operations because
ARM does pointer arithmetic at 65 bits, which caused macaw-symbolic to clobber
the block id in pointers. These corrupted pointers got fed to various parts of
the memory model, which caused trouble.

Note that this is on a macaw branch that needs to be merged.