Add a feature for comparing memory post-states of two binaries
This feature is intended to provide useful diagnostic information for the effects of patches in binaries that have undergone significant changes (e.g., replacing one function call with another that has different behavior).

This feature adds new configuration options (specified via the blockinfo files) to enable users to specify:
1. Pointers to memory regions that are expected to have different behavior
2. The addresses of two functions that are considered to be "morally equivalent' where the new analysis will run

When the verifier encounters the two functions (which are assumed to be relatively well-aligned, as the rest of the verifier requires), it will change modes. Instead of trying to break the program up into more tractable slices, it will treat the entire sub-trees of the call graph from that location as a unit and symbolically execute both the pre- and post-patch versions. The verifier then compares the post-states of memory after symbolic execution and summarizes the differences to the user. In particular, it reports the memory regions with different contents after symbolic execution except for:
- Regions explicitly ignored
- Dynamically-allocated memory that has been freed in the scope of the symbolic execution

This PR adds a new reporting feature to summarize interesting sub-proofs in a JSON format that is easier to consume than console output (`--proof-summary-json`). Note that this does not currently summarize all interesting nodes; it needs to be extended.

This PR also adds a command to capture the interaction with the SMT solver for debugging purposes (`--solver-interaction-file`).

Note that this PR currently has some dependencies on branches of macaw and crucible that need to be merged upstream after review (see #160 and #161).

Fixes #159 
Fixes #154
Fixes #133

Co-authored-by: Rob Dockins <rdockins@galois.com>