Revise the validity policy
There is now less implied code sharing, but this approach will reduce
complexity.  Now, the `ValidityPolicy` validates and returns the validated
term (whatever that means for the backend).  All of the logic for term creation
is now thus in the `ValidityPolicy`.

There are two policies right now:

1. The unsafe policy that just assumes that every operation works out
2. The "conservative" policy that injects undefined pointer terms (as undefined
   functions) as the original memory model did

This paves the way for the future optimistic memory model that generates
assertions that must be proved to safely use the simple terms, but would also
enable a fallback to the conservative terms when necessary.

The major improvements of this are:

1. Removing any direct dependency on the undefined pointers in the crucible
extension for macaw
2. Providing a new (and simpler) mechanism for identifying the undefined
operations that a term depends on (see notes in the code for the rationale and
function)