Be more explicit about handling of encoding and terminal properties
In the Docker image, the default encoding is unclear. The Dockerfile attempts to
set everything to UTF8, but it apparently does not succeed. Logging to a file in
the Docker image causes something in the GHC RTS to raise an encoding exception.

This change explicitly sets the encoding of the output file to UTF-8.  This
change is also more careful to not emit ANSI escapes when not outputting to a
terminal.