GHC 9 fixes (#183)
Update the codebase (and dependencies) to compile with ghc-9.0.2

Compiling with ghc-9 currently requires an `--allow-newer=base` because snap has not made a ghc-9 compatible release yet. The vast majority of the changes are eta expansions. Some of them are required because we have bundled constraints into the `EquivM` type alias, making most functions rank-n.