Simplify the logging interface
Now logging is separate from the interactive UI. This enables us to have a
single option for directing log output. Now the logging infrastructure hides the
fact that it uses channels, which simplifies the interface (and protocol for
setup/teardown).

There are separate channels for logging to a file and for streaming events to
the UI. These are combined at the logger level (i.e., if both are active, the
logger interface sends a message to *both* consumers, rather than having a
combined log consumer).

There is a change in default behavior: debug logs are never streamed to standard
output. The default is to not capture them at all, with an option (--log-file)
to direct them to a file.