Switch the PatchData config to TOML instead of a Read instance
The `Read` parser is very sensitive and fiddly, which is pretty painful to
users. The TOML parser is a lot more flexible and even allows comments in files,
which are useful for explaining what different values are for. It also enables
us to choose configuration keys with friendly names that are not valid Haskell
identifiers.

This also changes the structure of the `PatchData` type to be a bit more
self-describing.

Fixes #220