Add a configuration option to let users restrict the scope of verification
This change adds two options to the configuration file:
- `ignore-original-functions`
- `ignore-patched-functions`

These let the user specify a list of addresses that will be "ignored" in the
original (respectively patched) binary if they are functions that are called.
Operationally, a function is ignored by treating it as a no-op (and thus
ignoring all of its callees). This is unsafe, but can be useful for exploring
the impact of changes on large binaries.

Note that the while the two lists are independent, they probably need to align
very closely. They are separate to allow for minor code motion differences
between the two binaries, but if a function is ignored in one, it almost
certainly has to be ignored in the other, too.

Fixes #215