Improve github actions caching
This uses Ryan Scott's script to turn submodules into cabal-level source
repository dependencies in the cabal.project file. This means that the
submodules will be built into the cabal store instead of dist-newstyle, and thus
reused between runs.

This also rationalizes the cache keys, which were hashing a non-existent file.