Implement JSON reports for differential summaries
With this change, we have a concise way to report to users both differential
summaries and inlined memory differences without requiring users to trawl
through the text output.

Note that we could probably stand to improve the formatting of formulas (it is
currently just the printed output from what4).

Fixes #177