check equality for addresses in pre-domain
In the strongest postcondition verifier it is possible for
the incoming pre-domain to relax equality on locations that
the slice does not write to.

When computing the predicate to test for equality, we therefore
need to additionally check whether or not these locations are equal,
otherwise we will unsoundly omit them from the resulting domain
and implicitly assume that they are now equal.

fixes: #295