prefer concrete values when rewriting
Previously when applying an assumption set we would avoid
rewrites in the case of abiguity. This would result in cases
where augmenting an assumption frame
could suddenly cause symbolic terms to appear when they otherwise
would have been replaced with constant values.

Now we rewrite to the smallest constant in the target for a rewrite
(if it exists)