Add interactive debugger (#331)
This adds support for interacting with the verifier through a REPL that is implemented as a set of extensions to ghci. The `TraceTree` module defines most of the functionality for storing trace information in a tree structure, which has a dynamic layout according to the control flow of the verifier.

The `TraceTree` can support additional types by defining instances is the `IsTraceNode` class for a type-level string and defining the corresponding `TraceNodeType` type. 

The toplevel REPL is a relatively general interface for navigating a `TraceTree`: given a list of tracing nodes, the user can select a node by entering its corresponding menu number into the REPL, which then prints the contents of the node as a list of sub-nodes.

Also includes some verifier fixes:

- Fix issue during variable rescoping where the `maxRegion` field was not bound in the outgoing state, preventing it from being included in any equivalence domains.
- Add command-line flag for controlling the "error mode" for the verifier (i.e. how permissive error handling should be)