Allow desynchronization point to be provided when control flow desync is found
This simply modifies the disassembler to add a cut-point anywhere in
the function under analysis and then re-queues the node.
This allows the analysis to produce a valid (equivalent) CFAR just
before the control flow de-synchronization happens, so that the
domain from the one-sided analysis is as small as possible.