remove ExceptT transformer from EquivM
since EquivM functions are frequently unlifted into IO (via
MonadUnliftIO), errors that are raised in EquivM need to be thrown
as exceptions. Previously this resulted in some inconsistency in error handling
(i.e. 'finally' will only catch exceptions, but not errors).

This change unifies the EquivM errors to be thrown as exceptions, and
then just implements MonadError as a wrapper around these, just specialized
to the EquivalenceError type.