Pate.Script.withIOList: reduce delay when re-checking for new results
This fixes an issue with the Repl where the 'Choose Entry Point' prompt
would be shown before the script made an entry point selection, making
it appear that the verifier was blocked waiting for input.

Strictly speaking the delay in this function shouldn't be necessary, since
the execution is blocked on updates to the underlying list. However
it is possible for the continuation to update the underlying list
and therefore introduce a tight loop, which this delay would mitigate.