Move logic for starting two-sided analysis outside the diverging block exit
There is a widening step that happens when starting a one-sided analysis,
which connects the initial node to its Original variant. This ensures
that all of the domain information is carried forward from the two-sided
to the one-sided analysis.

Previously this widening step happened under the assumption that the
current path was the exit from the CFAR where the control flow divergence
was discovered. This assumption is too strong in general, since there
may be additional block exits in the two-sided analysis.