Add check for early termination of a block analysis
Handling a block exit may cause the current node to
require re-processing from the beginning (i.e. if
an assumption was added or desync/sync points were added).
This adds a check between each block exit for cases where
the current node is going to require re-analysis, ending
the analysis early if it does.

This is a small optimization, but also helps clean up the
output by reducing the number of redundant entries in the
Block Exits list.