Ensure that validInitState always runs with currentFunc set
The "currentFunc" field of the EquivalenceContext is initially
set to an "error" value until a function is under analysis.

This field is inspected by validInitState in order to resolve
architecture-specific validity constraints that are based on
the current function. (Specifically this is needed to
resolve the TOC on PPC).

This change adds a "withPair" wrapper around "validInitState" to
ensure that this field is always set in the EquivalenceContext
before it is given to any architecture-specific validity functions.