Add control flow alignment as strategy for handling diverging control flow
This adds an additional set of actions for adding an
assumption/assertion/equivalence condition stating that both
the original and patched CFARs execute the exact same instructions
in the same order.

This is intended to handle the common case where the analysis has
proceeded past the location of the patch and is now discovering
control flow divergences that are a result of the equivalence domain
being weakened from exact equality. These divergences are either
spurious and can be resolved by asserting control flow alignment, or
are actual differences introduced by the patch and can be resolved by
adding control flow alignment as an equivalence condition at this point.