Ensure that the macaw abstract domain overrides are set up correctly when handling returns
When analyzing a function with macaw we first install an override for
its (macaw) abstract domain based on any abstract domain information
we've collected during verification. This may help code discovery in
general, but is specifically required for ARM to handle cases where
macaw has lost track of the thumb mode flag and we need to use the
pate domain to recover it.

Prior to this fix, if the macaw discovery cache was flushed
(i.e. due to adding some additional block exit) then when a return
was processed, re-performing code discovery at the call site would not
use pate's domain information. In the case of ARM, this could cause
the disassembler to use the wrong mode.