Make symbol table lookups agnostic of low vs. high bit
In ARM the low bit of the function address indicates whether or not
it should be decoded in Thumb or ARM mode. However it is also possible
for the T processor flag to decide which mode to decode in. As a result
it is often possible for functions to be accessed via both addresses
(i.e. with the low bit set or unset).

As a workaround, we simply check both addresses when attempting to
resolve an address to a function symbol.