add support for symbolic-length writes by supporting reading array "chunks"
NOTE: the array chunks are represented by EventTrace.MemChunk, which
contains a source array, then base + length. Equality of chunks should
be possible via What4.Interface.arrayRangeEq but this seems to be
broken when using yices. As a workaround we compare equality of memchunks
by just looking at a concrete-length prefix, the (symbolic) length of the
chunk, and the last element in the chunk.