allow tests to terminate if the given script fails to match the output
the "script runner" thread now checks whether or not the current node
is blocked (waiting for user input) before re-executing the wait loop.
This allows us to detect when the script cannot make progress (i.e.
the verifier is waiting for a decision but we don't have matching input
to provide).

In the interactive case we don't necessarily need to fail or even report
this, but for tests/CI we need to consider this as "test failure"
and exit in order to avoid waiting indefinitely