wait until all choices have been added before blocking the TraceTree
This fixes an issue where scripts would fail to execute occasionally if
the script runner attempts to resolve a script step at exactly between
the points where the TraceTree has been blocked but before any choices
have been added.

In this case it would fail to find a match, but exits the script with
"failure" rather than re-trying because the TraceTree has "blocked" status,
indicating that it is waiting for input.