add SymSequenceTree datatype
this datatype drops the symbolic predicates inside a SymSequence while
retaining its internal branching structure. It is morally equivalent
to replacing all of the predicates in a SymSequence with arbitrary/free
variables.

the motivation for this is to represent all possible instruction paths
through a CFAR without consideration for the actual path conditions.
This representation can then be easily serialized to JSON and sent
to the GUI.