change 'no-assume-stack-scope' to 'assume-stack-scope' and add documentation
This flips the default behavior of this unsafe flag.

If set, 'assume-stack-scope' injects an additional
assumption into every pre-domain that all out of
scope stack values (i.e. past the stack pointer)

This is a fairly strong assumption that is unsafe in
general, in particular it will crash in the case where
it is provably false (i.e. the value domain implies that
the stack contents past the stack pointer are necessarily
inequivalent)

updates challenge10.args to use --assume-stack-scope