overhaul the control flow sync/merge logic to use uninterpreted functions
this gives a much more precise (and sound) treatment to how everything
is stitched together when combining two single-sided analysis nodes
into a two-sided node. Importantly, the "no-op" bundle for each
single sided analysis is modeled as making a no-op for the single-sided
node under consideration, but an undefined transition for the
other side of the analysis.

These undefined transitions are collected as uninterpreted functions, which
are eventually to be defined by propagating them backwards through
the single-sided analysis. Since this step is not complete, we now
expect to generate unprovable assertions which contain these functions.